+++ glyph/CHANGELOG.md
@@ -1,8 +1,13 @@
-# v0.2.16
+# Unreleased (0.2.17)
+* Add `VariableFont` trait implemented by `FontRef` & `FontVec`.
+  Provides `variations` & `set_variation` functions.
+* Add default enabled feature `variable-fonts`.
+
+# 0.2.16
 * Add `Font::pt_to_px_scale` to ease converting point size to `PxScale`.
 * Add `PxScale::round`.
 
-# v0.2.15
+# 0.2.15
 * Fix some font outlines by always trying to "close" them at the end. Fixes _Cantarell-VF.otf_ outlining.
 
 # 0.2.14